pyproject.toml
[project]
name = "documentapp"
version = "0.1.0"
description = ""
authors = [
    {name = "angelo-volpe", email = "angelovolpe95@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "Django>=5.1.2,<6.0.0",
    "djangorestframework>=3.15.2,<4.0.0",
    "drf-spectacular>=0.28.0,<0.29.0",
    "pillow>=11.0.0,<12.0.0",
    "psycopg2>=2.9.10,<3.0.0",
    "requests>=2.32.3,<3.0.0",
    "numpy>=2.2.1,<3.0.0",
    "opencv-python>=4.10.0.84,<5.0.0",
    "shapely>=2.0.6,<3.0.0",
    "pandas>=2.2.3,<3.0.0",
]

[dependency-groups]
dev = [
    "ruff>=0.8.0,<1.0.0",
    "pytest>=8.3.4,<9.0.0",
    "mypy>=1.0,<2.0",
    "coverage>=7.6.10,<8.0.0",
    "sphinx>=8.2.1,<9.0.0",
    "pre-commit>=4.5.1,<5.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
# Set the target Python version
target-version = "py312"

# Exclude common directories
exclude = [
    ".git",
    ".venv",
    "venv",
    "__pycache__",
    "build",
    "dist",
    "*.egg-info",
    ".mypy_cache",
    ".pytest_cache",
    ".ruff_cache",
    "migrations",
]
